Changes reqd to compile in VS2005
git-svn-id: svn://136.177.114.72/svn_GW/phreeqc/trunk@933 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ src/mainsubs.c
@@ -55,16 +55,16 @@ void initialize(void)
 	itmax=100;
 #ifdef USE_LONG_LDBLE	
 	/* from float.h, sets tolerance for cl1 routine */
-	ineq_tol = pow(10, -LDBL_DIG);
+	ineq_tol = pow((long double)10, (long double)-LDBL_DIG);
 #else
- 	ineq_tol = pow(10, -DBL_DIG);
+ 	ineq_tol = pow((double)10, (double)-DBL_DIG);
 #endif
 	convergence_tolerance = 1e-8;
 #ifdef USE_LONG_LDBLE	
 	/* from float.h, sets tolerance for cl1 routine */
-	inv_tol_default = pow(10, -LDBL_DIG+5);
+	inv_tol_default = pow((long double)10, (long double)-LDBL_DIG+5);
 #else
-	inv_tol_default = pow(10, -DBL_DIG+5);
+	inv_tol_default = pow((double)10, (double)-DBL_DIG+5);
 #endif
 	step_size = 100.;
 	pe_step_size = 10.;
